@@ -1,18 +1,20 @@
 package com.one.framework.app.login
 
-import android.content.Intent
+import android.app.Activity
 import android.graphics.Color
 import android.os.Bundle
 import android.text.Editable
+import android.text.Html
 import android.text.TextUtils
 import android.text.TextWatcher
 import android.view.View
-import com.one.framework.MainActivity
+import android.widget.Toast
 import com.one.framework.R
 import com.one.framework.app.base.BaseActivity
 import com.one.framework.log.Logger
+import com.one.framework.utils.ToastUtils
+import com.one.framework.utils.UIUtils
 import kotlinx.android.synthetic.main.one_login_layout.*
-import java.util.regex.Pattern
 
 class LoginActivity : BaseActivity(), View.OnClickListener {
     private lateinit var loginProxy: LoginProxy
@@ -22,9 +24,10 @@ class LoginActivity : BaseActivity(), View.OnClickListener {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.one_login_layout)
         with(oneLoginTitle) {
+            setLeftClickListener { skip() }
             hideRightImage(true)
-            setLeftImage(R.drawable.one_top_bar_back_normal)
-            setRightResId(R.string.one_login_skip)
+            setLeftImage(R.drawable.one_top_bar_back_selector)
+//            setRightResId(R.string.one_login_skip)
             setRightClickListener {
                 skip()
             }
@@ -34,33 +37,40 @@ class LoginActivity : BaseActivity(), View.OnClickListener {
         loginProxy = LoginProxy(this)
         oneLoginPhone.addTextChangedListener(watcher)
         oneLoginCode.addTextChangedListener(watcher)
+        oneLoginInvite.addTextChangedListener(watcher)
         oneLoginBtn.setRippleColor(Color.parseColor("#f3f3f3"), Color.parseColor("#f3f3f3"))
         oneLoginBtn.isEnabled = false
+
+        oneLoginPhone.setOnFocusChangeListener { _, hasFocus -> if (!hasFocus) checkPhone(oneLoginPhone.text.toString())  }
+        oneLoginProto.setText(Html.fromHtml(getString(R.string.one_login_proto)))
+        oneLoginProto.setOnClickListener(this)
+        parseIntent()
     }
 
     private fun skip() {
-        startActivity(Intent(this, MainActivity::class.java))
+//        startActivity(Intent(this, MainActivity::class.java))
         finish()
     }
 
-    private fun isMobileNO(phone: String): Boolean {
-        val telRegex = "^((13[0-9])|(14[5,7,9])|(15[^4])|(16[0-9])|(18[0-9])|(17[0,1,3,5,6,7,8]))\\d{8}$"
-        var pattern = Pattern.compile(telRegex)
-        return if (TextUtils.isEmpty(phone)) {
-            false
-        } else {
-            pattern.matcher(phone).matches()
-        }
+    private fun parseIntent() {
     }
 
     override fun onClick(v: View?) {
+        if (UIUtils.isFastDoubleClick()) {
+            return
+        }
         var id = v?.id
         when (id) {
             R.id.oneObtainCode -> {
-                doSendSms()
+                if (checkPhone(oneLoginPhone.text.toString())) {
+                    doSendSms()
+                }
             }
             R.id.oneLoginBtn -> {
                 doLogin()
+            }
+            R.id.oneLoginProto -> {
+
             }
         }
     }
@@ -68,12 +78,14 @@ class LoginActivity : BaseActivity(), View.OnClickListener {
 
     private fun doSendSms() {
         loginProxy.doSms(oneLoginPhone.text.toString(), object : ILogin.ILoginVerifyCode {
-            override fun onSuccess() {
+            override fun onSuccess(isNew: Boolean) {
+                oneInviteCodeLayout.visibility = if (isNew) View.VISIBLE else View.GONE
                 loginProxy.countDown(object : ILogin.ILoginCountDownTimer {
                     override fun onTick(millisUntilFinished: Long) {
-                        oneObtainCode.text = String.format(getString(R.string.one_login_recode_count_down), millisUntilFinished / 1000)
                         oneObtainCode.setTextColor(Color.parseColor("#999ba1"))
+                        oneObtainCode.text = UIUtils.highlight(String.format(getString(R.string.one_login_recode_count_down), millisUntilFinished / 1000), Color.parseColor("#1665ff"))
                     }
+
 
                     override fun onFinish() {
                         oneObtainCode.text = getString(R.string.one_login_verify_code)
@@ -94,18 +106,29 @@ class LoginActivity : BaseActivity(), View.OnClickListener {
         if (TextUtils.isEmpty(phone) && TextUtils.isEmpty(code)) {
             return
         }
-        loginProxy.doLogin(phone, code, object : ILogin.ILoginListener {
+        var inviteCode = oneLoginInvite.text.toString()
+        oneLoginLoading.visibility = View.VISIBLE
+        oneLoginBtn.tripButtonText = ""
+        loginProxy.doLogin(phone, code, inviteCode, object : ILogin.ILoginListener {
             override fun onLoginSuccess() {
+                setResult(Activity.RESULT_OK)
                 skip()
             }
 
             override fun onLoginFail(message: String?) {
+                setResult(-2)
+                ToastUtils.toast(this@LoginActivity, message)
+                oneLoginLoading.visibility = View.GONE
+                oneLoginBtn.setTripButtonText(R.string.one_login)
             }
         })
     }
 
-    fun checkPhone(phone: String) {
-        oneLoginCheck.visibility = if (!isMobileNO(phone)) View.VISIBLE else View.GONE
+    fun checkPhone(phone: String) : Boolean {
+        oneLoginCheck.visibility = if (UIUtils.isMobileNO(phone)) View.GONE else View.VISIBLE
+        oneObtainCode.isEnabled = UIUtils.isMobileNO(phone)
+        oneObtainCode.setTextColor(if (UIUtils.isMobileNO(phone)) Color.parseColor("#1665ff") else Color.parseColor("#babfca"))
+        return UIUtils.isMobileNO(phone)
     }
 
     inner class EditWatcher : TextWatcher {
@@ -116,6 +139,7 @@ class LoginActivity : BaseActivity(), View.OnClickListener {
             } else {
                 oneLoginCheck.visibility = View.GONE
             }
+            checkLoginStatus()
         }
 
         override fun beforeTextChanged(s: CharSequence?, start: Int, count: Int, after: Int) {
@@ -124,10 +148,14 @@ class LoginActivity : BaseActivity(), View.OnClickListener {
         }
 
         override fun onTextChanged(s: CharSequence?, start: Int, before: Int, count: Int) {
+
         }
     }
 
     private fun checkLoginStatus() {
-        oneLoginBtn.isEnabled = !TextUtils.isEmpty(oneLoginPhone.text.toString()) && !TextUtils.isEmpty(oneLoginCode.text.toString()) && isMobileNO(oneLoginPhone.text.toString())
+        var flag = (oneInviteCodeLayout.visibility == View.VISIBLE && !TextUtils.isEmpty(oneLoginInvite.text.toString())) || oneInviteCodeLayout.visibility == View.GONE
+        oneLoginBtn.isEnabled = !TextUtils.isEmpty(oneLoginPhone.text.toString())
+                && !TextUtils.isEmpty(oneLoginCode.text.toString())
+                && UIUtils.isMobileNO(oneLoginPhone.text.toString()) && flag
     }
 }
