@@ -2,19 +2,19 @@ package com.one.framework.app.service;
 
 import android.content.Context;
 import android.content.Intent;
+import android.os.Handler;
 import android.os.Message;
+import android.text.TextUtils;
 import com.igexin.sdk.GTIntentService;
-import com.igexin.sdk.PushConsts;
 import com.igexin.sdk.PushManager;
-import com.igexin.sdk.message.BindAliasCmdMessage;
-import com.igexin.sdk.message.FeedbackCmdMessage;
 import com.igexin.sdk.message.GTCmdMessage;
 import com.igexin.sdk.message.GTNotificationMessage;
 import com.igexin.sdk.message.GTTransmitMessage;
-import com.igexin.sdk.message.SetTagCmdMessage;
-import com.igexin.sdk.message.UnBindAliasCmdMessage;
-import com.one.framework.R;
+import com.one.framework.app.login.UserProfile;
+import com.one.framework.app.notify.NotifyManager;
 import com.one.framework.log.Logger;
+import com.one.framework.push.PushMessageManager;
+import com.one.framework.push.PushMsgModel;
 
 /**
  * Created by ludexiang on 2018/7/3.
@@ -28,13 +28,15 @@ import com.one.framework.log.Logger;
 public class GetuiIntentService extends GTIntentService {
 
   private static final String TAG = GetuiIntentService.class.getSimpleName();
+  private static Handler target;
 
-  /**
-   * 为了观察透传数据变化.
-   */
-  private static int cnt;
+  private int notifyId = 0x999;
 
   public GetuiIntentService() {
+  }
+
+  public void setHandler(Handler handler) {
+    target = handler;
   }
 
   @Override
@@ -64,55 +66,41 @@ public class GetuiIntentService extends GTIntentService {
     } else {
       String data = new String(payload);
       Logger.d(TAG, "receiver payload = " + data);
+      PushMessageManager pushMessageManager = new PushMessageManager(context, data);
+      PushMsgModel msgModel = pushMessageManager.getPushModel();
+      pushMessageManager.handlePush(msgModel);
+      sendMessage(msgModel);
 
-      // 测试消息为了观察数据变化
-      if (data.equals(getResources().getString(R.string.push_transmission_data))) {
-        data = data + "-" + cnt;
-        cnt++;
-      }
-      sendMessage(data, 0);
+      /** 通知栏展示 */
+      Intent intent = NotifyManager.getManager(context).buildIntent(context, msgModel);
+      NotifyManager.getManager(context).showNotification(msgModel.getTitle(), msgModel.getDescription(), msgModel.getMsgId(), intent);
     }
-
-    Logger.d(TAG, "----------------------------------------------------------------------------------------------");
   }
 
   @Override
   public void onReceiveClientId(Context context, String clientId) {
-    Logger.e(TAG, "onReceiveClientId -> " + "clientid = " + clientId);
-
-    sendMessage(clientId, 1);
-  }
-
-  @Override
-  public void onReceiveOnlineState(Context context, boolean online) {
-    Logger.d(TAG, "onReceiveOnlineState -> " + (online ? "online" : "offline"));
-  }
-
-  @Override
-  public void onReceiveCommandResult(Context context, GTCmdMessage cmdMessage) {
-    Logger.d(TAG, "onReceiveCommandResult -> " + cmdMessage);
-
-    int action = cmdMessage.getAction();
-
-    if (action == PushConsts.SET_TAG_RESULT) {
-      setTagResult((SetTagCmdMessage) cmdMessage);
-    } else if(action == PushConsts.BIND_ALIAS_RESULT) {
-      bindAliasResult((BindAliasCmdMessage) cmdMessage);
-    } else if (action == PushConsts.UNBIND_ALIAS_RESULT) {
-      unbindAliasResult((UnBindAliasCmdMessage) cmdMessage);
-    } else if ((action == PushConsts.THIRDPART_FEEDBACK)) {
-      feedbackResult((FeedbackCmdMessage) cmdMessage);
+    if (!TextUtils.isEmpty(clientId)) {
+      UserProfile.getInstance(context).savePushToken(clientId);
     }
   }
 
   @Override
-  public void onNotificationMessageArrived(Context context, GTNotificationMessage message) {
-//    NotifyManager.getManager(context).buildNotification("您有一条新消息", message.getTitle(), message.getContent(), new Intent());
+  public void onReceiveOnlineState(Context context, boolean online) {
+  }
 
+  @Override
+  public void onReceiveCommandResult(Context context, GTCmdMessage cmdMessage) {
+//    int action = cmdMessage.getAction();
+  }
+
+  @Override
+  public void onNotificationMessageArrived(Context context, GTNotificationMessage message) {
     Logger.d(TAG, "onNotificationMessageArrived -> " + "appid = " + message.getAppid() + "\ntaskid = " + message.getTaskId() + "\nmessageid = "
         + message.getMessageId() + "\npkg = " + message.getPkgName() + "\ncid = " + message.getClientId() + "\ntitle = "
         + message.getTitle() + "\ncontent = " + message.getContent());
 
+    /** 通知栏展示 */
+    NotifyManager.getManager(context).showNotification(message.getTitle(), message.getContent(), notifyId, new Intent());
   }
 
   @Override
@@ -122,160 +110,10 @@ public class GetuiIntentService extends GTIntentService {
         + message.getTitle() + "\ncontent = " + message.getContent());
   }
 
-  private void setTagResult(SetTagCmdMessage setTagCmdMsg) {
-    String sn = setTagCmdMsg.getSn();
-    String code = setTagCmdMsg.getCode();
-
-    int text = R.string.add_tag_unknown_exception;
-    switch (Integer.valueOf(code)) {
-      case PushConsts.SETTAG_SUCCESS:
-        text = R.string.add_tag_success;
-        break;
-
-      case PushConsts.SETTAG_ERROR_COUNT:
-        text = R.string.add_tag_error_count;
-        break;
-
-      case PushConsts.SETTAG_ERROR_FREQUENCY:
-        text = R.string.add_tag_error_frequency;
-        break;
-
-      case PushConsts.SETTAG_ERROR_REPEAT:
-        text = R.string.add_tag_error_repeat;
-        break;
-
-      case PushConsts.SETTAG_ERROR_UNBIND:
-        text = R.string.add_tag_error_unbind;
-        break;
-
-      case PushConsts.SETTAG_ERROR_EXCEPTION:
-        text = R.string.add_tag_unknown_exception;
-        break;
-
-      case PushConsts.SETTAG_ERROR_NULL:
-        text = R.string.add_tag_error_null;
-        break;
-
-      case PushConsts.SETTAG_NOTONLINE:
-        text = R.string.add_tag_error_not_online;
-        break;
-
-      case PushConsts.SETTAG_IN_BLACKLIST:
-        text = R.string.add_tag_error_black_list;
-        break;
-
-      case PushConsts.SETTAG_NUM_EXCEED:
-        text = R.string.add_tag_error_exceed;
-        break;
-
-      default:
-        break;
-    }
-
-    Logger.d(TAG, "settag result sn = " + sn + ", code = " + code + ", text = " + getResources().getString(text));
-  }
-
-  private void bindAliasResult(BindAliasCmdMessage bindAliasCmdMessage) {
-    String sn = bindAliasCmdMessage.getSn();
-    String code = bindAliasCmdMessage.getCode();
-
-    int text = R.string.bind_alias_unknown_exception;
-    switch (Integer.valueOf(code)) {
-      case PushConsts.BIND_ALIAS_SUCCESS:
-        text = R.string.bind_alias_success;
-        break;
-      case PushConsts.ALIAS_ERROR_FREQUENCY:
-        text = R.string.bind_alias_error_frequency;
-        break;
-      case PushConsts.ALIAS_OPERATE_PARAM_ERROR:
-        text = R.string.bind_alias_error_param_error;
-        break;
-      case PushConsts.ALIAS_REQUEST_FILTER:
-        text = R.string.bind_alias_error_request_filter;
-        break;
-      case PushConsts.ALIAS_OPERATE_ALIAS_FAILED:
-        text = R.string.bind_alias_unknown_exception;
-        break;
-      case PushConsts.ALIAS_CID_LOST:
-        text = R.string.bind_alias_error_cid_lost;
-        break;
-      case PushConsts.ALIAS_CONNECT_LOST:
-        text = R.string.bind_alias_error_connect_lost;
-        break;
-      case PushConsts.ALIAS_INVALID:
-        text = R.string.bind_alias_error_alias_invalid;
-        break;
-      case PushConsts.ALIAS_SN_INVALID:
-        text = R.string.bind_alias_error_sn_invalid;
-        break;
-      default:
-        break;
-
-    }
-
-    Logger.d(TAG, "bindAlias result sn = " + sn + ", code = " + code + ", text = " + getResources().getString(text));
-
-  }
-
-  private void unbindAliasResult(UnBindAliasCmdMessage unBindAliasCmdMessage) {
-    String sn = unBindAliasCmdMessage.getSn();
-    String code = unBindAliasCmdMessage.getCode();
-
-    int text = R.string.unbind_alias_unknown_exception;
-    switch (Integer.valueOf(code)) {
-      case PushConsts.UNBIND_ALIAS_SUCCESS:
-        text = R.string.unbind_alias_success;
-        break;
-      case PushConsts.ALIAS_ERROR_FREQUENCY:
-        text = R.string.unbind_alias_error_frequency;
-        break;
-      case PushConsts.ALIAS_OPERATE_PARAM_ERROR:
-        text = R.string.unbind_alias_error_param_error;
-        break;
-      case PushConsts.ALIAS_REQUEST_FILTER:
-        text = R.string.unbind_alias_error_request_filter;
-        break;
-      case PushConsts.ALIAS_OPERATE_ALIAS_FAILED:
-        text = R.string.unbind_alias_unknown_exception;
-        break;
-      case PushConsts.ALIAS_CID_LOST:
-        text = R.string.unbind_alias_error_cid_lost;
-        break;
-      case PushConsts.ALIAS_CONNECT_LOST:
-        text = R.string.unbind_alias_error_connect_lost;
-        break;
-      case PushConsts.ALIAS_INVALID:
-        text = R.string.unbind_alias_error_alias_invalid;
-        break;
-      case PushConsts.ALIAS_SN_INVALID:
-        text = R.string.unbind_alias_error_sn_invalid;
-        break;
-      default:
-        break;
-
-    }
-
-    Logger.d(TAG, "unbindAlias result sn = " + sn + ", code = " + code + ", text = " + getResources().getString(text));
-
-  }
-
-
-  private void feedbackResult(FeedbackCmdMessage feedbackCmdMsg) {
-    String appid = feedbackCmdMsg.getAppid();
-    String taskid = feedbackCmdMsg.getTaskId();
-    String actionid = feedbackCmdMsg.getActionId();
-    String result = feedbackCmdMsg.getResult();
-    long timestamp = feedbackCmdMsg.getTimeStamp();
-    String cid = feedbackCmdMsg.getClientId();
-
-    Logger.d(TAG, "onReceiveCommandResult -> " + "appid = " + appid + "\ntaskid = " + taskid + "\nactionid = " + actionid + "\nresult = " + result
-        + "\ncid = " + cid + "\ntimestamp = " + timestamp);
-  }
-
-  private void sendMessage(String data, int what) {
-    Message msg = Message.obtain();
-    msg.what = what;
-    msg.obj = data;
-//    DemoApplication.sendMessage(msg);
+  private void sendMessage(PushMsgModel model) {
+    Message msg = target.obtainMessage();
+    msg.what = model.getMsgId();
+    msg.obj = model;
+    msg.sendToTarget();
   }
 }
