@@ -6,17 +6,20 @@ import android.content.ClipData;
 import android.content.Intent;
 import android.net.Uri;
 import android.os.Build;
+import android.os.Build.VERSION;
 import android.provider.MediaStore;
+import android.support.v4.content.FileProvider;
 import android.view.View;
 import android.webkit.ValueCallback;
 import com.one.framework.R;
 import com.one.framework.app.web.jsbridge.functions.image.BottomListMenu;
+import com.one.framework.app.web.jsbridge.functions.image.BottomListMenu.OnDismissListener;
+import com.one.framework.log.Logger;
 import com.one.framework.utils.FileUtils;
 import java.io.File;
 
 /**
  * WebActivity 文件选择管理类
- * Created by zhengtao on 17/8/3.
  */
 
 public class FileChooserManager {
@@ -41,6 +44,8 @@ public class FileChooserManager {
    */
   private ValueCallback<Uri[]> uploadMessageAboveL;
 
+  private BottomListMenu mAvatarMenu;
+
 
   public FileChooserManager(WebActivity webActivity) {
     mActivity = webActivity;
@@ -55,7 +60,6 @@ public class FileChooserManager {
   public void onActivityResult(int requestCode, int resultCode, Intent data) {
     /* 文件(图片)选择 */
     if (requestCode == REQUEST_CODE_SELECT_PIC || requestCode == REQUEST_CODE_CAPTURE_PIC) {
-
       if (uploadMessageAboveL != null) {
         handleFileChooseResultAboveL(requestCode, resultCode, data);
       } else {
@@ -118,10 +122,10 @@ public class FileChooserManager {
       if (mChosenFile != null) {
         results = new Uri[]{Uri.fromFile(mChosenFile)};
       }
-
     }
     uploadMessageAboveL.onReceiveValue(results);
     uploadMessageAboveL = null;
+    mChosenFile = null;
   }
 
 
@@ -133,43 +137,64 @@ public class FileChooserManager {
       showMenuDialog();
     }
 
-
+    /**
+     * above L invoke this method
+     * @param filePathCallback
+     */
     @Override
     public void openFileChooserAboveL(ValueCallback<Uri[]> filePathCallback) {
       uploadMessageAboveL = filePathCallback;
       showMenuDialog();
     }
-
-    private void showMenuDialog() {
-      BottomListMenu mAvatarMenu = new BottomListMenu(mActivity, mRootView,
-          mActivity.getResources().getStringArray(R.array.one_photos));
-      mAvatarMenu.setListMenuListener(new BottomListMenu.ListMenuListener() {
-        @Override
-        public void onItemSelected(int position, String itemStr) {
-          if (position == 0) {
-            Intent intent = new Intent(MediaStore.ACTION_IMAGE_CAPTURE);
-            mChosenFile = FileUtils.getPhotoOutputFile();
-            intent.putExtra(MediaStore.EXTRA_OUTPUT, Uri.fromFile(mChosenFile));
-            mActivity.startActivityForResult(intent, REQUEST_CODE_CAPTURE_PIC);
-          } else if (position == 1) {
-            Intent intent = new Intent();
-            intent.setType("image/*");
-            intent.setAction(Intent.ACTION_GET_CONTENT);
-            try {
-              mActivity.startActivityForResult(intent, REQUEST_CODE_SELECT_PIC);
-            } catch (Exception e) {
-            }
-          }
-        }
-      });
-      mAvatarMenu.setDismissListener(new BottomListMenu.OnDismissListener() {
-        @Override
-        public void dismiss() {
-          mChosenFile = null;
-          onActivityResult(REQUEST_CODE_CAPTURE_PIC, Activity.RESULT_OK, null);
-        }
-      });
-      mAvatarMenu.showDialog();
-    }
   };
+
+  /**
+   * native choose image method
+   */
+  private void showMenuDialog() {
+    mAvatarMenu = new BottomListMenu(mActivity, mRootView, mActivity.getResources().getStringArray(R.array.one_photos));
+    mAvatarMenu.setListMenuListener((position, itemStr) -> {
+      if (position == 0) {
+        Intent intent = new Intent(MediaStore.ACTION_IMAGE_CAPTURE);
+        mChosenFile = FileUtils.getPhotoOutputFile(mActivity.getPackageName());
+        intent.addCategory(Intent.CATEGORY_DEFAULT);
+        Uri uri;
+        if (VERSION.SDK_INT >= Build.VERSION_CODES.N){
+          uri = FileProvider.getUriForFile(mActivity.getApplicationContext(), "com.one.trip.FileProvider", mChosenFile);
+        } else {
+          uri = Uri.fromFile(mChosenFile);
+        }
+        Logger.e("ldx",  "uri >> " + uri);
+        intent.putExtra(MediaStore.EXTRA_OUTPUT, uri);
+        intent.setFlags(Intent.FLAG_GRANT_READ_URI_PERMISSION);
+        mActivity.startActivityForResult(intent, REQUEST_CODE_CAPTURE_PIC);
+      } else if (position == 1) {
+        Intent intent = new Intent();
+        intent.setAction(Intent.ACTION_GET_CONTENT);
+        intent.setType("image/*");
+        try {
+          mActivity.startActivityForResult(intent, REQUEST_CODE_SELECT_PIC);
+        } catch (Exception e) {
+        }
+      }
+    });
+    mAvatarMenu.setDismissListener(new OnDismissListener() {
+      @Override
+      public void dismiss() {
+        mChosenFile = null;
+        onActivityResult(REQUEST_CODE_SELECT_PIC, Activity.RESULT_CANCELED, null);
+      }
+    });
+    mAvatarMenu.showDialog();
+  }
+
+  public void dismiss() {
+    if (mAvatarMenu != null && mAvatarMenu.isShowing()) {
+      mAvatarMenu.dismiss();
+    }
+  }
+
+  public boolean isShowing() {
+    return mAvatarMenu != null && mAvatarMenu.isShowing();
+  }
 }
