@@ -1,6 +1,7 @@
 package com.one.framework.app.widget;
 
 import android.content.Context;
+import android.content.res.TypedArray;
 import android.graphics.Color;
 import android.graphics.Typeface;
 import android.text.TextUtils;
@@ -29,6 +30,7 @@ public class ListItemWithRightArrowLayout extends RelativeLayout implements ILis
   private ScriptTextView mRightTxt;
   private IClickCallback mClickCallback;
   private ImageView mArrow;
+  private View mSeparator;
 
   public ListItemWithRightArrowLayout(Context context) {
     this(context, null);
@@ -43,23 +45,51 @@ public class ListItemWithRightArrowLayout extends RelativeLayout implements ILis
     mContext = context;
 
     View view = LayoutInflater.from(context).inflate(R.layout.one_list_item_with_arrow_layout, this, true);
-    mItemImg = (ShapeImageView) view.findViewById(R.id.list_item_img_view);
-    mItemTitle = (TextView) view.findViewById(R.id.list_item_info);
-    mRightTxt = (ScriptTextView) view.findViewById(R.id.list_item_right_info);
+    mItemImg = view.findViewById(R.id.list_item_img_view);
+    mItemTitle = view.findViewById(R.id.list_item_info);
+    mRightTxt = view.findViewById(R.id.list_item_right_info);
     mArrow = view.findViewById(R.id.list_item_right_arrow);
+    mSeparator = view.findViewById(R.id.list_item_separator);
 
     setClickable(true);
     setBackgroundDrawable(UIUtils.rippleDrawableRect(Color.WHITE, Color.parseColor("#e3e3e3")));
+
+    TypedArray a = context.obtainStyledAttributes(attrs, R.styleable.ListItemWithRightArrowLayout);
+    int itemTitle = a.getResourceId(R.styleable.ListItemWithRightArrowLayout_itemTitle, -1);
+    int itemColor = a.getColor(R.styleable.ListItemWithRightArrowLayout_itemTitleColor, -1);
+    int itemSize = a.getDimensionPixelSize(R.styleable.ListItemWithRightArrowLayout_itemTitleSize, -1);
+    int itemMargin = a.getDimensionPixelOffset(R.styleable.ListItemWithRightArrowLayout_itemMargin, -1);
+    boolean showArrow = a.getBoolean(R.styleable.ListItemWithRightArrowLayout_itemShowArrow, true);
+    String itemRightTitle = a.getString(R.styleable.ListItemWithRightArrowLayout_itemRightTitle);
+    int itemRightUnit = a.getResourceId(R.styleable.ListItemWithRightArrowLayout_itemRightUnit, -1);
+    int itemRightUnitColor = a.getColor(R.styleable.ListItemWithRightArrowLayout_itemRightUnitColor, Color.parseColor("#999ba1"));
+    int itemRightUnitSize = a.getInt(R.styleable.ListItemWithRightArrowLayout_itemRightUnitSize, -1);
+    int itemRightColor = a.getColor(R.styleable.ListItemWithRightArrowLayout_itemRightTitleColor, Color.parseColor("#999ba1"));
+    a.recycle();
+
+    if (itemTitle != -1 && itemColor != -1 && itemSize != -1 && itemMargin != -1) {
+      setItemTitle(itemTitle, itemSize, itemColor, false);
+      setLRMargin(itemMargin);
+    }
+
+    if (!TextUtils.isEmpty(itemRightTitle)) {
+      setRightTxt(itemRightTitle, itemRightColor);
+    }
+    if (itemRightUnit != -1) {
+      setScriptTxt(getResources().getString(itemRightUnit), itemRightUnitColor, itemRightUnitSize);
+    }
+    setArrowVisible(showArrow);
   }
 
   @Override
   public void setLRMargin(int margin) {
-    RelativeLayout.LayoutParams params = (RelativeLayout.LayoutParams)mItemTitle.getLayoutParams();
+    RelativeLayout.LayoutParams params = (RelativeLayout.LayoutParams) mItemTitle.getLayoutParams();
     params.leftMargin = params.rightMargin = margin;
     mItemTitle.setLayoutParams(params);
 
-    RelativeLayout.LayoutParams arrowParams = (RelativeLayout.LayoutParams) mArrow.getLayoutParams();
-    arrowParams.leftMargin = arrowParams.rightMargin = margin;
+    RelativeLayout.LayoutParams arrowParams = (RelativeLayout.LayoutParams) mArrow
+        .getLayoutParams();
+    arrowParams.rightMargin = margin;
     mArrow.setLayoutParams(arrowParams);
   }
 
@@ -100,6 +130,12 @@ public class ListItemWithRightArrowLayout extends RelativeLayout implements ILis
   }
 
   @Override
+  public void setItemTitle(int strRes, int color) {
+    mItemTitle.setText(strRes);
+    mItemTitle.setTextColor(color);
+  }
+
+  @Override
   public void setItemTitle(CharSequence title) {
     mItemTitle.setText(title);
   }
@@ -112,8 +148,22 @@ public class ListItemWithRightArrowLayout extends RelativeLayout implements ILis
 
   @Override
   public void setItemTitle(CharSequence title, int size, int color) {
+    setItemTitle(title, size, color, true);
+  }
+
+  @Override
+  public void setItemTitle(int title, int size, int color, boolean useSPUnit) {
+    setItemTitle(getResources().getString(title), size, color, useSPUnit);
+  }
+
+  @Override
+  public void setItemTitle(CharSequence title, int size, int color, boolean useSPUnit) {
     mItemTitle.setText(title);
-    mItemTitle.setTextSize(TypedValue.COMPLEX_UNIT_SP, size);
+    if (useSPUnit) {
+      mItemTitle.setTextSize(TypedValue.COMPLEX_UNIT_SP, size);
+    } else {
+      mItemTitle.setTextSize(TypedValue.COMPLEX_UNIT_PX, size);
+    }
     mItemTitle.setTextColor(color);
   }
 
@@ -121,7 +171,8 @@ public class ListItemWithRightArrowLayout extends RelativeLayout implements ILis
   public void setItemTitle(CharSequence title, int size, boolean bold, int color) {
     mItemTitle.setText(title);
     mItemTitle.setTextSize(TypedValue.COMPLEX_UNIT_SP, size);
-    mItemTitle.setTypeface(bold ? Typeface.defaultFromStyle(Typeface.BOLD) : Typeface.defaultFromStyle(Typeface.NORMAL));
+    mItemTitle.setTypeface(bold ? Typeface.defaultFromStyle(Typeface.BOLD)
+        : Typeface.defaultFromStyle(Typeface.NORMAL));
     mItemTitle.setTextColor(color);
   }
 
@@ -133,6 +184,16 @@ public class ListItemWithRightArrowLayout extends RelativeLayout implements ILis
   @Override
   public void setRightTxt(int resId) {
     mRightTxt.setText(resId);
+  }
+
+  @Override
+  public void setItemColor(int color) {
+    mItemTitle.setTextColor(color);
+  }
+
+  @Override
+  public void setRightColor(int color) {
+    mRightTxt.setTextColor(color);
   }
 
   @Override
@@ -151,6 +212,24 @@ public class ListItemWithRightArrowLayout extends RelativeLayout implements ILis
   @Override
   public void setScriptTxt(CharSequence scriptTxt, int color, int size) {
     mRightTxt.setScriptText((String) scriptTxt, color, size);
+  }
+
+  /**
+   * 设置separator color
+   */
+  @Override
+  public void setSeparatorColor(int color) {
+
+  }
+
+  @Override
+  public void setArrowVisible(boolean visible) {
+    mArrow.setVisibility(visible ? View.VISIBLE : View.GONE);
+    if (!visible) {
+      RelativeLayout.LayoutParams params = (RelativeLayout.LayoutParams) mRightTxt.getLayoutParams();
+      params.addRule(RelativeLayout.ALIGN_PARENT_RIGHT);
+      mRightTxt.setLayoutParams(params);
+    }
   }
 
   @Override
