@@ -3,7 +3,6 @@ package com.one.framework.app.widget;
 import android.animation.Animator;
 import android.animation.AnimatorListenerAdapter;
 import android.animation.ValueAnimator;
-import android.animation.ValueAnimator.AnimatorUpdateListener;
 import android.annotation.TargetApi;
 import android.content.Context;
 import android.content.res.TypedArray;
@@ -12,6 +11,7 @@ import android.util.AttributeSet;
 import android.util.SparseArray;
 import android.view.LayoutInflater;
 import android.view.View;
+import android.view.animation.LinearInterpolator;
 import android.widget.AbsListView;
 import android.widget.AbsListView.OnScrollListener;
 import android.widget.AdapterView;
@@ -35,7 +35,7 @@ public class PullListView extends ListView implements IMovePublishListener, IPul
   private int mCurrentFirstVisibleItem = 0;
 
   private IHeaderView mHeaderView;
-  private int mScroller; // 0 scroll Header 1 scroll self
+  private int mScrollerType; // 0 scroll Header 1 scroll self
   private int mMaxHeight;
 
   private boolean isHaveFooterView;
@@ -45,8 +45,13 @@ public class PullListView extends ListView implements IMovePublishListener, IPul
   private int mRefreshLayoutId;
   private boolean isHaveLoadMoreView;
   private int mLoadMoreLayoutId;
+  private LoadMoreLayout mLoadMoreLayout;
   private IItemClickListener mItemClickListener;
   private IPullCallback mPullListener;
+  private boolean isLoading;
+  private int mTotalCount;
+  private ILoadListener mListener;
+  private boolean hasMore;
 
   private int upTranslateY;
 
@@ -62,7 +67,7 @@ public class PullListView extends ListView implements IMovePublishListener, IPul
     super(context, attrs, defStyleAttr);
 
     TypedArray a = context.obtainStyledAttributes(attrs, R.styleable.PullView);
-    mScroller = a.getInt(R.styleable.PullView_scroll_view, 1);
+    mScrollerType = a.getInt(R.styleable.PullView_scroll_view, 1);
     mMaxHeight = a.getDimensionPixelSize(R.styleable.PullView_scroll_max_height, 0);
 
     isHaveFooterView = a.getBoolean(R.styleable.PullView_have_footer_view, true);
@@ -74,7 +79,7 @@ public class PullListView extends ListView implements IMovePublishListener, IPul
     mLoadMoreLayoutId = a.getResourceId(R.styleable.PullView_load_more_view_id, -1);
     a.recycle();
 
-    if (mScroller == 0 && mMaxHeight == 0) {
+    if (mScrollerType == 0 && mMaxHeight == 0) {
       throw new IllegalArgumentException("ScrollMaxHeight is 0");
     }
 
@@ -84,13 +89,13 @@ public class PullListView extends ListView implements IMovePublishListener, IPul
 
     if (isHaveLoadMoreView && mLoadMoreLayoutId == -1) {
       throw new IllegalArgumentException("Please set LoadMoreViewLayout");
+    } else {
+      inflateFooterView();
     }
 
     mHeaderView = new HeaderView(getContext(), mMaxHeight);
     setOnScrollListener(this);
     setOnItemClickListener(this);
-
-    inflate();
   }
 
   @Override
@@ -119,7 +124,7 @@ public class PullListView extends ListView implements IMovePublishListener, IPul
   @Override
   public void addFooterView(View v) {
     if (isHaveFooterView) {
-      super.addFooterView(v);
+      super.addFooterView(v, null, false);
     }
   }
 
@@ -129,20 +134,15 @@ public class PullListView extends ListView implements IMovePublishListener, IPul
   }
 
   @Override
-  public void setHeaderView(int layout) {
-    mHeaderView.setHeaderView(layout);
-  }
-
-  private void inflate() {
-    if (mRefreshLayoutId != -1) {
-      addHeaderView(LayoutInflater.from(getContext()).inflate(mRefreshLayoutId, null));
-    }
+  public void setHeaderView(int headerViewHeight) {
+    mHeaderView.setHeaderView(headerViewHeight);
   }
 
   @Override
   protected void onMeasure(int widthMeasureSpec, int heightMeasureSpec) {
     if (isResolveConflict) {
-      super.onMeasure(widthMeasureSpec, MeasureSpec.makeMeasureSpec(Integer.MAX_VALUE >> 2, MeasureSpec.AT_MOST));
+      super.onMeasure(widthMeasureSpec,
+          MeasureSpec.makeMeasureSpec(Integer.MAX_VALUE >> 2, MeasureSpec.AT_MOST));
     } else {
       super.onMeasure(widthMeasureSpec, heightMeasureSpec);
     }
@@ -150,13 +150,22 @@ public class PullListView extends ListView implements IMovePublishListener, IPul
 
   @Override
   public void onScrollStateChanged(AbsListView view, int scrollState) {
-
+    int lastVisibleIndex = getLastVisiblePosition();
+    if (!isLoading && scrollState == OnScrollListener.SCROLL_STATE_IDLE//停止滚动
+        && lastVisibleIndex == mTotalCount - 1) {//滑动到最后一项
+      isLoading = true;
+      if (mListener != null && !hasMore) {
+        loadFootView();
+        mListener.onLoadMore();
+      }
+    }
   }
 
   @Override
   public void onScroll(AbsListView view, int firstVisibleItem, int visibleItemCount,
       int totalItemCount) {
     mCurrentFirstVisibleItem = firstVisibleItem;
+    mTotalCount = totalItemCount;
     View firstView = view.getChildAt(0);
     if (null != firstView) {
       ItemRecord itemRecord = recordSp.get(firstVisibleItem);
@@ -172,6 +181,7 @@ public class PullListView extends ListView implements IMovePublishListener, IPul
   }
 
   class ItemRecord {
+
     int height = 0;
     int top = 0;
   }
@@ -207,8 +217,10 @@ public class PullListView extends ListView implements IMovePublishListener, IPul
 
   @Override
   public boolean isScrollBottom() {
-    if (getLastVisiblePosition() != INVALID_POSITION && getLastVisiblePosition() == (getCount() - 1)) {
-      final View bottomChildView = getChildAt(getLastVisiblePosition() - getFirstVisiblePosition());
+    if (getLastVisiblePosition() != INVALID_POSITION && getLastVisiblePosition() == (getCount()
+        - 1)) {
+      final View bottomChildView = getChildAt(
+          getLastVisiblePosition() - getFirstVisiblePosition());
       return getHeight() >= bottomChildView.getBottom();
     }
     return false;
@@ -217,59 +229,63 @@ public class PullListView extends ListView implements IMovePublishListener, IPul
   @Override
   public void onMove(float offsetX, float offsetY) {
     boolean flag = offsetY > 0 || isHeaderNeedScroll();
-    if (mScroller == 0 && flag && !isScrollBottom()) {
+    if (mScrollerType == 0 && flag && !isScrollBottom()) {
       mHeaderView.onMove(offsetX, offsetY);
     } else {
-      selfScrollerMove(offsetX, offsetY);
+      selfScrollerMove(offsetX, (int) (offsetY + 0.5f));
     }
   }
 
   @Override
   public void onUp(boolean bottom2Up, boolean isFling) {
-    if (mScroller == 0 && isHeaderNeedScroll()) { // 如果是滚到到底部了则滚动listview
+    if (mScrollerType == 0 && isHeaderNeedScroll()) { // 如果是滚到到底部了则滚动listview
       mHeaderView.onUp(bottom2Up, isFling);
     } else {
       selfScrollerUp(bottom2Up, isFling);
     }
   }
 
-  private void selfScrollerMove(float offsetX, float offsetY) {
-    int translateY = (int) (getTranslationY() + offsetY + 0.5);
-    setTranslationY(translateY);
+  /**
+   * @param offsetY Y轴上的偏移量
+   */
+  private void selfScrollerMove(float offsetX, int offsetY) {
+    int translateY = (int) (getTranslationY() + offsetY);
     if (mPullListener != null) {
-     mPullListener.move(offsetX, translateY);
+      mPullListener.move(offsetX, offsetY);
     }
+    setTranslationY(translateY);
   }
 
   private void selfScrollerUp(boolean bottom2Up, boolean isFling) {
-    int tranlationY = upTranslateY = (int) getTranslationY();
-    goonMove(200);
+    int translationY = upTranslateY = (int) getTranslationY();
+    goonMove(ANIM_DURATION, false);
   }
 
-  private void goonMove(long duration) {
+  @Override
+  public void goonMove(long duration, boolean rollback) {
     ValueAnimator translate = ValueAnimator.ofFloat(1f, 0f);
     translate.setDuration(duration);
-    translate.addUpdateListener(new AnimatorUpdateListener() {
-      @Override
-      public void onAnimationUpdate(ValueAnimator animation) {
-        float animValue = animation.getAnimatedFraction();
-        float fraction = 1f - animValue;
-        float translationY = fraction * getTranslationY();
-        setTranslationY(translationY);
-        if (mPullListener != null) { // 没满的时候自动回缩
-          mPullListener.move(0, translationY);
-        }
+    translate.addUpdateListener(animation -> {
+      float animValue = animation.getAnimatedFraction(); // 百分比 0 ~ 1
+      float fraction = 1f - animValue; // value 1 ~ 0
+      float translationY =
+          (rollback || isScrollBottom() ? 0 : animValue * mCustomHeaderViewHeight)
+              + fraction * getTranslationY();
+      if (mPullListener != null && !rollback) { // 没满的时候自动回缩
+        mPullListener.moveAfterUp(0, fraction);
       }
+      setTranslationY(translationY);
     });
     translate.addListener(new AnimatorListenerAdapter() {
       @Override
       public void onAnimationEnd(Animator animation) {
         super.onAnimationEnd(animation);
-        if (mPullListener != null) {
-          mPullListener.up(upTranslateY);
+        if (mPullListener != null && !rollback) {
+          mPullListener.up(getTranslationY());
         }
       }
     });
+    translate.setInterpolator(new LinearInterpolator());
     translate.start();
   }
 
@@ -298,5 +314,77 @@ public class PullListView extends ListView implements IMovePublishListener, IPul
   @Override
   public void setPullCallback(IPullCallback listener) {
     mPullListener = listener;
+  }
+
+  @Override
+  public void setLoadListener(ILoadListener listener) {
+    mListener = listener;
+  }
+
+  /******************** 自定义 headerView *********************/
+  private int mCustomHeaderViewHeight;
+
+  /**
+   * 设置刷新View
+   */
+  @Deprecated
+  private void inflateRefreshView(int layoutId) {
+    View refreshView = null;
+    if (mRefreshLayoutId != -1) {
+      refreshView = LayoutInflater.from(getContext()).inflate(mRefreshLayoutId, null);
+    }
+//    inflateRefreshView();
+    View.inflate(getContext(), layoutId, null);
+    addHeaderView(refreshView);
+//    if (!isScrollBottom()) {
+//      /**
+//       * 默认刷新view不展示，根据收拾下滑展示刷新view scrollTo 向上向左滑动为正 向下向右为负
+//       */
+//      scrollTo(0, mRefreshView.getRefreshHeight());
+//    }
+  }
+
+  @Override
+  public void setCustomHeaderViewHeight(int viewHeight) {
+    mCustomHeaderViewHeight = viewHeight;
+  }
+
+  /****************** 自定义 headerView *********************/
+
+  private void inflateFooterView() {
+    View footerView = null;
+    if (mLoadMoreLayoutId != -1) {
+      footerView = LayoutInflater.from(getContext()).inflate(mLoadMoreLayoutId, null);
+      mLoadMoreLayout = footerView.findViewById(R.id.one_load_more);
+    }
+    if (footerView != null) {
+//    inflateRefreshView();
+      addFooterView(footerView);
+    }
+//    if (!isScrollBottom()) {
+//      /**
+//       * 默认刷新view不展示，根据收拾下滑展示刷新view scrollTo 向上向左滑动为正 向下向右为负
+//       */
+//      scrollTo(0, mRefreshView.getRefreshHeight());
+//    }
+  }
+
+  public void loadFootView() {
+    if (mLoadMoreLayout != null) {
+      mLoadMoreLayout.loading();
+    }
+  }
+
+  public void releaseFootView(boolean noMore) {
+    hasMore = noMore;
+    isLoading = false;
+    if (mLoadMoreLayout != null) {
+      mLoadMoreLayout.noMore(noMore);
+    }
+  }
+
+  public void canLoadMore() {
+    hasMore = false;
+    isLoading = false;
   }
 }
