@@ -4,14 +4,17 @@ import static android.view.MotionEvent.INVALID_POINTER_ID;
 
 import android.content.Context;
 import android.support.v4.view.MotionEventCompat;
+import android.support.v4.view.ViewCompat;
 import android.util.AttributeSet;
 import android.view.MotionEvent;
 import android.view.VelocityTracker;
 import android.view.View;
 import android.view.ViewConfiguration;
+import android.widget.AbsListView;
 import android.widget.RelativeLayout;
 import com.one.framework.app.widget.base.IMovePublishListener;
 import com.one.framework.app.widget.base.IPullView;
+import com.one.framework.log.Logger;
 import com.one.framework.utils.SystemUtils;
 import com.one.framework.utils.UIUtils;
 
@@ -20,6 +23,8 @@ import com.one.framework.utils.UIUtils;
  */
 
 public class PullScrollRelativeLayout extends RelativeLayout {
+
+  private static final String TAG = PullScrollRelativeLayout.class.getSimpleName();
 
   private static float RATIO = 3.0f;
   private Context mContext;
@@ -30,6 +35,7 @@ public class PullScrollRelativeLayout extends RelativeLayout {
   private int mMaxVelocity;
   private int mLastDownX, mLastDownY;
   private int mActionDownX, mActionDownY;
+  private int mTouchDownX, mTouchDonwY;
   private View mScrollView;
   private IPullView mPullView;
   private boolean isScrolling = false;
@@ -104,6 +110,7 @@ public class PullScrollRelativeLayout extends RelativeLayout {
           mLastDownY = mActionDownY = (int) y;
           // Save the ID of this pointer (for dragging)
           mActivePointerId = MotionEventCompat.getPointerId(ev, 0);
+          Logger.i("Pull", "downX >>>> " + mLastDownY  + " downY " + mLastDownY);
         }
         break;
       }
@@ -115,7 +122,7 @@ public class PullScrollRelativeLayout extends RelativeLayout {
         final int curX = (int) MotionEventCompat.getX(ev, pointerIndex);
         final int curY = (int) MotionEventCompat.getY(ev, pointerIndex);
         boolean isScrollTop = (curY - mLastDownY >= mMinScroll && mPullView.getScrollingPadding() == 0) || (mScrollView.getTranslationY() > 0 && !mPullView.isScrollBottom()) || mPullView.isHeaderNeedScroll();
-        boolean isScrollBottom = (mLastDownY - curY >= mMinScroll && mPullView.isScrollBottom() && mScrollView.getTranslationY() <= 0);
+        boolean isScrollBottom = (mLastDownY - curY >= mMinScroll && mPullView.isScrollBottom()) && mScrollView.getTranslationY() <= 0;
         isScrolling = isScrollTop || isScrollBottom;
         if (checkScrollView()) {
           if (canScroll()) {
@@ -128,18 +135,23 @@ public class PullScrollRelativeLayout extends RelativeLayout {
             handleMove(offsetX, offsetY);
             mActionDownX = curX;
             mActionDownY = curY;
-//            return false; // 去掉此处之后 有滑动阻尼的时候 listItem 就不会出现按下的效果了
+//            return true; // 去掉此处之后 有滑动阻尼的时候 listItem 就不会出现按下的效果了
           } else {
-            // 不是顶部 | 底部的时候设置mScrollView的Y轴方向滚动为0
+            // 不是顶部 | 底部的时候设置mScrollView的Y轴方向滚动为0 so far 简单粗暴
             mScrollView.setTranslationY(0);
+//            handleMove(0, 0);
           }
         }
+
         break;
       }
       case MotionEvent.ACTION_CANCEL:
       case MotionEvent.ACTION_OUTSIDE:
       case MotionEvent.ACTION_UP: {
+        int upActionX = (int) ev.getRawX();
+        int upActionY = (int) ev.getRawY();
         int scrollY = (int) mScrollView.getTranslationY();
+        Logger.v("Pull", "up mLastDownX >>>> " + mLastDownX + " DownY " + mLastDownY + " upX " + upActionX + " upY " + upActionY + " scrollY " + scrollY);
         if (scrollY != 0) { // 有阻尼效果
           mTracker.computeCurrentVelocity(1000, mMaxVelocity);
           float yVelocity = mTracker.getYVelocity();
@@ -155,9 +167,23 @@ public class PullScrollRelativeLayout extends RelativeLayout {
     return super.dispatchTouchEvent(ev);
   }
 
+  /**
+   * 这段代码来自谷歌官方的 SwipeRefreshLayout
+   * 应用场景已经在英文注释中解释清楚
+   * 大部分第三方下拉刷新库都保留了这段代码，本库也不例外
+   */
   @Override
-  public boolean onTouchEvent(MotionEvent event) {
-    return true;
+  public void requestDisallowInterceptTouchEvent(boolean disallowIntercept) {
+    // if this is a List < L or another view that doesn't support nested
+    // scrolling, ignore this request so that the vertical scroll event
+    // isn't stolen
+    View target = mPullView.getView();
+    if ((android.os.Build.VERSION.SDK_INT >= 21 || !(target instanceof AbsListView))
+        && (target == null || ViewCompat.isNestedScrollingEnabled(target))) {
+      super.requestDisallowInterceptTouchEvent(disallowIntercept);
+      //} else {
+      // Nope.
+    }
   }
 
   private void handleMove(float offsetX, float offsetY) {
