@@ -78,7 +78,12 @@ public class PullScrollView extends ScrollView implements IPullView, IMovePublis
   }
 
   @Override
-  public void setHeaderView(int layout) {
+  public void setHeaderView(int headerViewHeight) {
+
+  }
+
+  @Override
+  public void setCustomHeaderViewHeight(int layout) {
 
   }
 
@@ -141,10 +146,11 @@ public class PullScrollView extends ScrollView implements IPullView, IMovePublis
 
   private void selfScrollerUp(boolean bottom2Up, boolean isFling) {
     int tranlationY = (int) getTranslationY();
-    goonMove(200);
+    goonMove(ANIM_DURATION, false);
   }
 
-  private void goonMove(long duration) {
+  @Override
+  public void goonMove(long duration, boolean rollback) {
     ValueAnimator translate = ValueAnimator.ofFloat(1f, 0f);
     translate.setDuration(duration);
     translate.addUpdateListener(new AnimatorUpdateListener() {
@@ -175,6 +181,11 @@ public class PullScrollView extends ScrollView implements IPullView, IMovePublis
 
   @Override
   public void setPullCallback(IPullCallback listener) {
+
+  }
+
+  @Override
+  public void setLoadListener(ILoadListener listener) {
 
   }
 }
