@@ -13,7 +13,8 @@ public class WebpageQueryHelper {
   private static boolean isHostQualified(@NonNull Uri uri) {
     // taxi must be fix m  / o/ b
     return !TextUtils.isEmpty(uri.getHost()) &&
-        (uri.getHost().endsWith("taxi.com") || uri.getHost().endsWith("taxi.io"));
+        (uri.getHost().endsWith("taxi.com") || uri.getHost().endsWith("taxi.io")
+            || uri.getHost().endsWith("mobike.com") || uri.getHost().endsWith("mobike.io"));
   }
 
 
